Added the Feature and PerceptronImage classes
A Feature points to 4 random pixels and starts with a random boolean.
You can call evaluate on the feature with a PerceptronImage as a
parameter and it will check the 4 pixels on the image. If 3 or more of
the pixels are what the feature expects it will return true, otherwise
false.

The PerceptronImage holds a series of boolean values (stored in a 2d
array) that represent a black and white image. It is also passed in
whether it is an x or not.